#SELFIE - visual: add pound/hash signs to tags
IT IS NOT CALLED HASHTAG! "#" is a hash or pound sign, or sharp if you're musical.
People are familiar with using hash # to tag things; let's make them feel at home.

KNOWN BUG: This should be optional, but when I tried to make it dependent on a SiteParam, Hugo didn't like it; some sort of context error. It's also missing from the Tags overview page, but tags need fixing anyway.