[Feature] Add "Last Modified" SiteParam to footer
Feature is optional parameter, useful to serve as a website's "heartbeat", so the reader does not have to search through all the articles to see if the site is still alive.

[BUG] Does not display an accurate date. No idea why or where it pulls the date from, but you can e.g. have a site with published articles dated later than the "Last Updated" date. Spooky.

Later, we may want to use .GitInfo.Hash (https://discourse.gohugo.io/t/question-about-implementation/36499) to link "Last Modified" with the latest commit, but it was complicated to set up (https://stackoverflow.com/questions/57141806/how-can-i-get-gitinfo-in-hugo), requiring Git binaries in PATH and stuff, so for now, a configurable URL for build logs (or RSS feed, or anything) will serve.